move token test to correct directory
